refactor components into separate files
