+++ src/components/ProfilePicker.tsx
@@ -1,23 +1,7 @@
-import {
-  Avatar,
-  AvatarFallback,
-  AvatarImage,
-  avatarVariants,
-} from "@/components/ui/avatar";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import {
-  AutomergeUrl,
-  DocHandle,
-  DocumentId,
-  Repo,
-  isValidAutomergeUrl,
-  stringifyAutomergeUrl,
-} from "@automerge/automerge-repo";
-import { useDocument, useRepo } from "@automerge/automerge-repo-react-hooks";
-import { EventEmitter } from "eventemitter3";
-import { Copy, Eye, EyeOff, User as UserIcon } from "lucide-react";
+import { ContactDoc, ProfileDoc, useProfile, useSelf } from "@/profile";
+import { DocumentId, stringifyAutomergeUrl } from "@automerge/automerge-repo";
+import { isValidDocumentId } from "@automerge/automerge-repo/dist/AutomergeUrl"; // todo don't import this
+import { ChangeEvent, useEffect, useState } from "react";
 
 import {
   Dialog,
@@ -26,6 +10,15 @@ import {
   DialogHeader,
   DialogTrigger,
 } from "@/components/ui/dialog";
+
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+
+import { useDocument } from "@automerge/automerge-repo-react-hooks";
+
+import { Copy, Eye, EyeOff } from "lucide-react";
+
 import { Label } from "@/components/ui/label";
 import {
   Tooltip,
@@ -33,288 +26,7 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
-import { type VariantProps } from "class-variance-authority";
-import { ChangeEvent, useEffect, useMemo, useReducer, useState } from "react";
-import { isValidDocumentId } from "@automerge/automerge-repo/dist/AutomergeUrl";
-
-interface ProfileDoc {
-  contactUrl: AutomergeUrl;
-}
-
-interface AnonymousContactDoc {
-  type: "anonymous";
-  claimedBy?: AutomergeUrl;
-}
-
-interface RegisteredContactDoc {
-  type: "registered";
-  name: string;
-  avatarUrl?: AutomergeUrl;
-}
-
-type ContactDoc = AnonymousContactDoc | RegisteredContactDoc;
-
-interface ProfileEvents {
-  change: () => void;
-}
-
-interface ContactProps {
-  name: string;
-  avatar: File;
-}
-
-class Profile extends EventEmitter<ProfileEvents> {
-  #repo: Repo;
-  #handle: DocHandle<ProfileDoc>;
-  #contactHandle: DocHandle<ContactDoc>;
-
-  constructor(
-    repo: Repo,
-    handle: DocHandle<ProfileDoc>,
-    contactHandle: DocHandle<ContactDoc>
-  ) {
-    super();
-
-    this.#repo = repo;
-    this.#handle = handle;
-    this.#contactHandle = contactHandle;
-
-    // listen for changed profileUrl caused by other tabs
-    window.addEventListener("storage", async (event) => {
-      if (event.key === PROFILE_URL) {
-        const newProfileUrl = event.newValue as AutomergeUrl;
-
-        // try to see if profile is already loaded
-        const profileHandle = this.#repo.find<ProfileDoc>(newProfileUrl);
-        const profileDoc = await profileHandle.doc();
-        if (profileDoc.contactUrl) {
-          this.logIn(newProfileUrl);
-          return;
-        }
-
-        // ... otherwise wait until contactUrl of profile is loaded
-        profileHandle.on("change", ({ doc }) => {
-          if (doc.contactUrl) {
-            this.logIn(newProfileUrl);
-          }
-        });
-      }
-    });
-  }
-
-  async logIn(profileUrl: AutomergeUrl) {
-    // override old profileUrl
-    localStorage.setItem(PROFILE_URL, profileUrl);
-
-    const profileHandle = this.#repo.find<ProfileDoc>(profileUrl);
-    const profileDoc = await profileHandle.doc();
-    const contactHandle = this.#repo.find<ContactDoc>(profileDoc.contactUrl);
-
-    this.#contactHandle.change((oldContact: AnonymousContactDoc) => {
-      if (oldContact.type === "anonymous") {
-        oldContact.claimedBy = contactHandle.url;
-      }
-    });
-
-    this.#contactHandle = contactHandle;
-    this.#handle = profileHandle;
-    this.emit("change");
-  }
-
-  async signUp({ name, avatar }: ContactProps) {
-    let avatarUrl: AutomergeUrl;
-    if (avatar) {
-      avatarUrl = await uploadFile(this.#repo, avatar);
-    }
-
-    this.contactHandle.change((contact: RegisteredContactDoc) => {
-      contact.type = "registered";
-      contact.name = name;
-
-      if (avatarUrl) {
-        contact.avatarUrl = avatarUrl;
-      }
-    });
-  }
-
-  async logOut() {
-    const profileHandle = this.#repo.create<ProfileDoc>();
-    const contactHandle = this.#repo.create<ContactDoc>();
-
-    profileHandle.change((profile) => {
-      profile.contactUrl = contactHandle.url;
-    });
-
-    contactHandle.change((contact) => {
-      contact.type = "anonymous";
-    });
-
-    localStorage.setItem(PROFILE_URL, profileHandle.url);
-
-    this.#handle = profileHandle;
-    this.#contactHandle = contactHandle;
-
-    this.emit("change");
-  }
-
-  get handle() {
-    return this.#handle;
-  }
-
-  get contactHandle() {
-    return this.#contactHandle;
-  }
-}
-
-const PROFILE_URL = "tinyEssayEditor:profileUrl";
-
-let CURRENT_PROFILE: Promise<Profile>;
-
-async function getProfile(repo: Repo) {
-  if (!repo.storageSubsystem) {
-    throw new Error("cannot create profile without storage");
-  }
-
-  if (CURRENT_PROFILE) {
-    const currentProfile = await CURRENT_PROFILE;
-    if (currentProfile) {
-      return currentProfile;
-    }
-  }
-
-  const profileUrl = localStorage.getItem(PROFILE_URL) as AutomergeUrl;
-
-  // try to load existing profile
-  if (profileUrl) {
-    CURRENT_PROFILE = new Promise<Profile>(async (resolve) => {
-      const profileHandle = repo.find<ProfileDoc>(profileUrl);
-      const contactHandle = repo.find<ContactDoc>(
-        (await profileHandle.doc()).contactUrl
-      );
-      resolve(new Profile(repo, profileHandle, contactHandle));
-    });
-
-    return CURRENT_PROFILE;
-  }
-
-  // ... otherwise create a new one
-  const profileHandle = repo.create<ProfileDoc>();
-  const contactHandle = repo.create<ContactDoc>();
-
-  profileHandle.change((profile) => {
-    profile.contactUrl = contactHandle.url;
-  });
-
-  contactHandle.change((contact) => {
-    contact.type = "anonymous";
-  });
-
-  localStorage.setItem(PROFILE_URL, profileHandle.url);
-  const newProfile = new Profile(repo, profileHandle, contactHandle);
-  CURRENT_PROFILE = Promise.resolve(newProfile);
-  return newProfile;
-}
-
-function useForceUpdate() {
-  const [, forceUpdate] = useReducer((x) => x + 1, 0);
-  return forceUpdate;
-}
-
-export function useProfile(): Profile | undefined {
-  const repo = useRepo();
-  const [profile, setProfile] = useState<Profile | undefined>(undefined);
-
-  const forceUpdate = useForceUpdate();
-
-  useEffect(() => {
-    getProfile(repo).then(setProfile);
-  }, [repo]);
-
-  useEffect(() => {
-    if (!profile) {
-      return;
-    }
-
-    profile.on("change", forceUpdate);
-
-    return () => {
-      profile.off("change", forceUpdate);
-    };
-  }, [profile]);
-
-  return profile;
-}
-
-function useProfileDoc(): ProfileDoc {
-  const profile = useProfile();
-  const [profileDoc] = useDocument<ProfileDoc>(profile?.handle.url);
-  return profileDoc;
-}
-
-export function useSelf(): ContactDoc {
-  const profileDoc = useProfileDoc();
-  const [contactDoc] = useDocument<ContactDoc>(profileDoc?.contactUrl);
-
-  return contactDoc;
-}
-
-const initials = (name: string) => {
-  return name
-    .split(" ")
-    .map((word) => word[0])
-    .join("");
-};
-
-interface ContactAvatarProps extends VariantProps<typeof avatarVariants> {
-  url?: AutomergeUrl;
-  showName?: boolean;
-  name?: string;
-  avatar?: File;
-}
-
-export const ContactAvatar = ({
-  name: nameOverride,
-  avatar: avatarOverride,
-  url,
-  showName = false,
-  size,
-}: ContactAvatarProps) => {
-  const [maybeAnonymousContact] = useDocument<ContactDoc>(url);
-  const [registeredContact] = useDocument<RegisteredContactDoc>(
-    maybeAnonymousContact?.type === "anonymous"
-      ? maybeAnonymousContact.claimedBy
-      : undefined
-  );
-
-  const contact: RegisteredContactDoc =
-    maybeAnonymousContact?.type === "registered"
-      ? maybeAnonymousContact
-      : registeredContact;
-
-  const avatarBlobUrl = useBlobUrl(contact?.avatarUrl);
-  const avatarOverrideUrl = useMemo(
-    () => avatarOverride && URL.createObjectURL(avatarOverride),
-    [avatarOverride]
-  );
-
-  const avatarUrl = avatarOverrideUrl
-    ? avatarOverrideUrl
-    : url && contact?.avatarUrl
-    ? avatarBlobUrl
-    : undefined;
-  const name = nameOverride ?? contact?.name;
-
-  return (
-    <div className="flex items-center gap-1.5">
-      <Avatar size={size}>
-        <AvatarImage src={avatarUrl} alt={name} />
-        <AvatarFallback>{name ? initials(name) : <UserIcon />}</AvatarFallback>
-      </Avatar>
-
-      {showName && <b>{name ?? "Anonymous"}</b>}
-    </div>
-  );
-};
+import { ContactAvatar } from "./ContactAvatar";
 
 enum ProfilePickerTab {
   LogIn = "logIn",
@@ -566,44 +278,4 @@ export const ProfilePicker = () => {
       </DialogContent>
     </Dialog>
   );
-};
-
-interface FileDoc {
-  type: string;
-  data: ArrayBuffer;
-}
-
-const uploadFile = async (repo: Repo, file: File): Promise<AutomergeUrl> => {
-  const reader = new FileReader();
-  const fileDocHandle = repo.create<FileDoc>();
-
-  const isLoaded = new Promise((resolve) => {
-    reader.onload = (event) => {
-      fileDocHandle.change((fileDoc) => {
-        fileDoc.type = file.type;
-        fileDoc.data = new Uint8Array(event.target.result as ArrayBuffer);
-      });
-
-      resolve(true);
-    };
-  });
-
-  reader.readAsArrayBuffer(file);
-
-  await isLoaded;
-  return fileDocHandle.url;
-};
-
-const useBlobUrl = (url: AutomergeUrl) => {
-  const [file] = useDocument<FileDoc>(url);
-
-  return useMemo(() => {
-    if (!file || !file.data || !file.type) {
-      return;
-    }
-
-    const blob = new Blob([file.data], { type: file.type });
-    const url = URL.createObjectURL(blob);
-    return url;
-  }, [file?.data, file?.type]);
 };
